Split test framework types into separate files
