+++ src/xunit.performance.execution/BenchmarkTestFrameworkTypeDiscoverer.cs
@@ -1,0 +1,20 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using Xunit.Abstractions;
+using Xunit.Sdk;
+
+namespace Microsoft.Xunit.Performance
+{
+    internal class BenchmarkTestFrameworkTypeDiscoverer : ITestFrameworkTypeDiscoverer
+    {
+        public Type GetTestFrameworkType(IAttributeInfo attribute)
+        {
+            if (BenchmarkConfiguration.RunningAsPerfTest)
+                return typeof(BenchmarkTestFramework);
+            else
+                return typeof(XunitTestFramework);
+        }
+    }
+}
